refactor(adr-018): delete ThreeSessionRunner; execution.ts calls runThreeSessionTddFromCtx directly
